+++ jsmetrics/metrics/jet_core_algorithms.py
@@ -30,7 +30,7 @@ def koch_et_al_2006(data, ws_threshold=30):
     The weighted average windspeed for the jet events is calculated as follows:
 
     .. math::
-        \alpha vel =  \frac{1}{(p2-p1)} \int_{p1}^{p2} (u^2+v^2)^{1/2} \,dp
+        \alpha vel =  \frac{1}{p2-p1} \int_{p1}^{p2} (u^2+v^2)^{1/2} \,dp
 
     where p1, p2 is min, max pressure level.
 
@@ -111,16 +111,20 @@ def koch_et_al_2006(data, ws_threshold=30):
 
 
 @sort_xarray_data_coords(coords=["lat", "lon"])
-def schiemann_et_al_2009(data):
+def schiemann_et_al_2009(data, ws_threshold=30):
     r"""
-    This method detects jet occurrences, whereby each occurence is detected based
-    on three rules applied to inputted wind speed (V, u, v):
-        1. \|V\| is a local maxima in longitude and altitude plane
-        2. \|V\| > 30 m/s
-        3. \|u\| > 0 m/s.
+    This method detects jet occurrences, whereby each jet occurence is detected based
+    on three rules applied to inputted wind speed (V = [u, v]):
+        1. \|V\| is a local maxima in latitude and altitude plane
+        2. \|V\| ≥ 30 m/s
+        3. \|u\| ≥ 0 m/s.
 
     This method was originally introduce in Schiemann et al 2009 (https://doi.org/10.1175/2008JCLI2625.1)
     and is described in Section 2 of that study.
+
+    Please see 'Notes' below for any additional information about the implementation of this method
+    to this package.
+
 
     Parameters
     ----------
@@ -138,44 +142,74 @@ def schiemann_et_al_2009(data):
     Notes
     -----
     While the original method is built on a four dimension slice of wind speed (time, lat, lon, plev),
-    This implementation will work where there is only one plev.
+    This implementation will work where there is only one pressure level, so a 3-d slice (time, lat, lon).
 
     **Slow method:** Due to the nature of this method, it currently takes a very long time to run,
     i.e. 8 seconds per time unit on AMD Ryzen 5 3600 6-core processor.
 
+    Examples
+    --------
+    .. code-block:: python
+
+        import jsmetrics
+        import xarray as xr
+
+        # Load in dataset with u and v components:
+        uv_data = xr.open_dataset('path_to_uv_data')
+
+        # Subset dataset to range used in original methodology (100-500 hPa & 16.7-58.25 N, 42.5-220.5 E)):
+        uv_sub = uv_data.sel(plev=slice(100, 500), lat=slice(16.7, 58.25), lon=slice(42.5, 220.5))
+
+        # Run algorithm:
+        schiemann_outputs = jsmetrics.jet_core_algorithms.schiemann_et_al_2009(uv_sub, ws_threshold=30)
+
+        # Produce a jet occurence count across all pressure levels
+        schiemann_jet_counts_all_levels = schiemann_outputs['jet_occurence'].sum(('time', 'plev'))
+
+        # Use the jet occurence values as a mask to extract the jet windspeeds
+        schiemann_jet_ws = schiemann_outputs.where(schiemann_outputs['jet_occurence'] > 0)['ws']
     """
     #  Step 1. Calculate wind vector
     data["ws"] = windspeed_utils.get_resultant_wind(data["ua"], data["va"])
 
-    #  Step 2. Calculate jet maximas
+    #  Step 2. Calculate jet occurences
     if "time" not in data.coords:
         raise KeyError("Please provide a time coordinate for data to run this metric")
     if data["time"].size == 1:
-        output = jet_core_algorithms_components.get_local_jet_maximas_by_oneday_by_plev(
-            data
+        if "time" in data.dims:
+            data = data.squeeze("time")
+        output = (
+            jet_core_algorithms_components.get_local_jet_occurence_by_oneday_by_plev(
+                data, ws_threshold=ws_threshold
+            )
         )
     else:
         output = data.groupby("time").map(
-            jet_core_algorithms_components.get_local_jet_maximas_by_oneday_by_plev
+            jet_core_algorithms_components.get_local_jet_occurence_by_oneday_by_plev,
+            (ws_threshold,),
         )
     return output
 
 
 @sort_xarray_data_coords(coords=["lat", "lon"])
 def manney_et_al_2011(data, ws_core_threshold=40, ws_boundary_threshold=30):
-    """
-    Looks to get seperate jet cores based on boundary and threshold. Core are discovered where 8-cells are above boundary threshold
-    Paper uses 100-400 hPa.
+    r"""
+    This method seperates jet cores based on boundary and windspeed threshold.
+    Core are discovered where 8-cells are above boundary threshold
 
-    Method from Manney et al. (2011) https://doi.org/10.5194/acp-11-6115-2011
-    Also see Manney et al. 2011, 2014, 2017 and Manney & Hegglin 2018
+    This method is also known as JETPAC and is available online (add link).
 
-    NOTE: Currently takes a long time i.e. 7.6 seconds per time unit with 8 plevs (i.e. 7.6 seconds per day) on AMD Ryzen 5 3600 6-core processor
+    This method was originally introduce in Manney et al. (2011) (https://doi.org/10.5194/acp-11-6115-2011),
+    and is described in Section 3.1 of that study.
+
+    There is an update to this method introduced in Manney & Hegglin 2018 to include physically based method to extract the
+    subtropical jet is identified (and thus distinguished from polar jets).
+
 
     Parameters
     ----------
     data : xarray.Dataset
-        Data containing u- and v-component wind
+        Data which should containing the variables: 'ua' and 'va', and the coordinates: 'lon', 'lat', 'plev' and 'time'.
     ws_core_threshold : int or float
         Threshold used for jet-stream core point (default=40)
     ws_boundary_threshold : int or float
@@ -184,7 +218,35 @@ def manney_et_al_2011(data, ws_core_threshold=40, ws_boundary_threshold=30):
     Returns
     ----------
     output : xarray.Dataset
-        Data containing jet-cores (ID number relates to each unique core)
+        Data containing the variable 'jet-core_id' (ID number relates to each unique core)
+
+    Notes
+    -----
+    **Slow method**: currently takes a long time i.e. 7.6 seconds per time unit with 8 plevs
+    (i.e. 7.6 seconds per day) on AMD Ryzen 5 3600 6-core processor
+
+    Examples
+    --------
+    .. code-block:: python
+
+        import jsmetrics
+        import xarray as xr
+
+        # Load in dataset with u and v components:
+        uv_data = xr.open_dataset('path_to_uv_data')
+
+        # Subset dataset to range used in original methodology (100-400 hPa)):
+        uv_sub = uv_data.sel(plev=slice(100, 400))
+
+        # Run algorithm:
+        manney_outputs = jsmetrics.jet_core_algorithms.manney_et_al_2011(uv_sub, ws_core_threshold=40, ws_boundary_threshold=30)
+
+        # Produce a jet occurence count across all pressure levels
+        manney_jet_counts_all_levels = manney_outputs['jet_core_id'].sum(('time', 'plev'))
+
+        # Use the jet occurence values as a mask to extract the jet windspeeds
+        manney_jet_ws = manney_outputs.where(manney_outputs['jet_core_id'] > 0)['ws']
+
     """
     if "plev" not in data.dims:
         data = data.expand_dims("plev")
@@ -194,7 +256,7 @@ def manney_et_al_2011(data, ws_core_threshold=40, ws_boundary_threshold=30):
         raise KeyError("Please provide a time coordinate for data to run this metric")
     if data["time"].size == 1:
         if "time" in data.dims:
-            data = data.isel(time=0)
+            data = data.squeeze("time")
         output = jet_core_algorithms_components.run_jet_core_algorithm_on_one_day(
             data, ws_core_threshold, ws_boundary_threshold
         )
@@ -211,22 +273,63 @@ def manney_et_al_2011(data, ws_core_threshold=40, ws_boundary_threshold=30):
 
 @sort_xarray_data_coords(coords=["lat", "lon"])
 def penaortiz_et_al_2013(data):
-    """
-    Will calculate local wind maxima days per monthyear
+    r"""
+    This method follows a two step procedure for calculate local wind maxima days.
+    This method returns 3 outputted variables:
+        1. local_wind_maxima
+        2. polar_front_jet
+        3. subtropical_jet
+    Each output is in a monthyear frequency.
 
-    Method from Pena-Ortiz (2013) https://doi.org/10.1002/jgrd.50305
+    This method was first introduced in Pena-Ortiz et al. (2013) (https://doi.org/10.1002/jgrd.50305) and
+    is described in section 2 of that study.
 
-    NOTE: Currently takes a long time i.e. 1.3 seconds per time unit with 8 plevs (i.e. 1.3 seconds per day) on AMD Ryzen 5 3600 6-core processor
+    Please see 'Notes' below for any additional information about the implementation of this method
+    to this package.
 
     Parameters
     ----------
     data : xarray.Dataset
-        Data containing u- and v-component wind
+        Data which should containing the variables: 'ua' and 'va', and the coordinates: 'lon', 'lat', 'plev' and 'time'.
 
     Returns
     ----------
     output : xarray.Dataset
         Data containing number of days per month with local wind maxima
+
+    Notes
+    -----
+    Currently takes a long time i.e. 1.3 seconds per time unit with 8 plevs (i.e. 1.3 seconds per day)
+    on AMD Ryzen 5 3600 6-core processor
+
+    Examples
+    --------
+    .. code-block:: python
+
+        import jsmetrics
+        import xarray as xr
+
+        # Load in dataset with u and v components:
+        uv_data = xr.open_dataset('path_to_uv_data')
+
+        # Subset dataset to range used in original methodology (100-400 hPa)):
+        uv_sub = uv_data.sel(plev=slice(100, 400))
+
+        # Run algorithm:
+        pena_outputs = jsmetrics.penaortiz_et_al_2013(uv_sub)
+
+        # Produce a count of polar front jet values across all pressure levels
+        pena_pfj_counts_all_levels = pena_outputs['polar_front_jet'].sum(('monthyear', 'plev'))
+
+        # Use the polar front jet values as a mask to extract the jet windspeeds
+        uv_sub["ws_by_monthyear"] = (
+            data["ws"]
+            .resample(time="MS")
+            .sum()
+            .rename({"time": "monthyear"})
+        )
+        pena_pfj_ws = pena_outputs.where(pena_outputs['polar_front_jet'] > 0)['ws_by_monthyear']
+
     """
     #  Step 1. Calculate wind vector
     data["ws"] = windspeed_utils.get_resultant_wind(data["ua"], data["va"])
@@ -257,24 +360,61 @@ def penaortiz_et_al_2013(data):
 
 @sort_xarray_data_coords(coords=["lat", "lon"])
 def kuang_et_al_2014(data, occurence_ws_threshold=30):
-    """
-    Looks to get event-based jet occurrence and jet center occurrence of JS (1 is occurence, 2 is core).
+    r"""
+    This method produces an event-based jet occurrence and jet center occurrence of JS.
+    The outputs of this method will produce categorical values of three types:
+        0. is not determined to be part of the jet
+        1. is a jet occurence
+        2. is jet core (upgraded from a jet occurence)
 
-    Method from Kuang et al (2014) https://doi.org/10.1007/s00704-013-0994-x
+    This method was first introduced in Kuang et al (2014) (https://doi.org/10.1007/s00704-013-0994-x) and
+    is described in section 2 of that study.
 
-    NOTE: Currently takes a long time i.e. 2 seconds per time unit with 1 plev (i.e. 2 seconds per day) on AMD Ryzen 5 3600 6-core processor
+    Please see 'Notes' below for any additional information about the implementation of this method
+    to this package.
+
 
     Parameters
     ----------
     data : xarray.Dataset
-        Data containing u- and v-component wind
+        Data which should containing the variables: 'ua' and 'va', and the coordinates: 'lon', 'lat', 'plev' and 'time'.
     occurence_ws_threshold : int or float
         Threshold used to identify a jet-stream occurence point (default=30)
 
     Returns
     ----------
     output : xarray.Dataset
-        Data containing jet-occurence and jet-centres (1 is occurence, 2 is core)
+        Data containing jet-occurence and jet-centres (1 is occurence, 2 is core, 0 is no jet)
+
+    Notes
+    -----
+    Currently takes a long time i.e. 2 seconds per time unit with 1 plev (i.e. 2 seconds per day) on AMD Ryzen 5 3600 6-core processor
+
+    Examples
+    --------
+    .. code-block:: python
+
+        import jsmetrics
+        import xarray as xr
+
+        # Load in dataset with u and v components:
+        uv_data = xr.open_dataset('path_to_uv_data')
+
+        # Subset dataset to range used in original methodology (200 hPa)):
+        uv_sub = uv_data.sel(plev=slice(200, 200))
+
+        # Run algorithm:
+        kuang_outputs = jsmetrics.kuang_et_al_2014(uv_sub, occurence_ws_threshold=30)
+
+        # Extract only jet centers
+        kuang_jet_centers = kuang_outputs.where(kuang_outputs['jet_ocurrence1_jet_centre2']==2)
+
+        # Produce a count of jet centers across all pressure levels
+        kuang_jet_centers_counts_all_levels = kuang_jet_centers.sum(('time', 'plev'))
+
+        # Use the jet occurence values as a mask to extract the jet windspeeds
+        kuang_jet_ws = kuang_outputs.where(kuang_outputs['jet_ocurrence1_jet_centre2'] > 0)['ws']
+
     """
     if "plev" in data.dims:
         if data["plev"].count() == 1:
@@ -291,6 +431,8 @@ def kuang_et_al_2014(data, occurence_ws_threshold=30):
         )
     else:
         if data["time"].size == 1:
+            if "time" in data.dims:
+                data = data.squeeze("time")
             output = jet_core_algorithms_components.run_jet_occurence_and_centre_alg_on_one_day(
                 data, occurence_ws_threshold
             )